# app/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Construction Dashboard API",
    "description": "API for Construction Project Management Dashboard",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/daily-reports/create": {
      "post": {
        "tags": [
          "Field"
        ],
        "summary": "Create a new daily report",
        "description": "Creates a new daily report for a project",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Daily report object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DailyReportCreate"
            }
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/DailyReportResponse"
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/daily-reports/{id}": {
      "get": {
        "tags": [
          "Field"
        ],
        "summary": "Get daily report by ID",
        "description": "Returns a specific daily report",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Daily Report ID",
            "required": true,
            "type": "integer"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/DailyReportDetail"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Daily report not found"
          }
        }
      }
    },
    "/projects/{id}/safety/observations": {
      "get": {
        "tags": [
          "Safety"
        ],
        "summary": "Get safety observations for project",
        "description": "Returns all safety observations for a specific project",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "type": "integer"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/SafetyObservationList"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Project not found"
          }
        }
      }
    },
    "/projects/{id}/safety/incidents": {
      "get": {
        "tags": [
          "Safety"
        ],
        "summary": "Get safety incidents for project",
        "description": "Returns all safety incidents for a specific project",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "type": "integer"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/IncidentList"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Project not found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DailyReportDetail": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "data": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "report_number": {
              "type": "string"
            },
            "report_date": {
              "type": "string",
              "format": "date"
            },
            "weather_conditions": {
              "type": "string"
            },
            "temperature_high": {
              "type": "number"
            },
            "temperature_low": {
              "type": "number"
            },
            "precipitation": {
              "type": "number"
            },
            "wind_speed": {
              "type": "number"
            },
            "manpower_count": {
              "type": "integer"
            },
            "delays": {
              "type": "boolean"
            },
            "delay_description": {
              "type": "string"
            },
            "work_performed": {
              "type": "string"
            },
            "materials_received": {
              "type": "string"
            },
            "equipment_used": {
              "type": "string"
            },
            "visitors": {
              "type": "string"
            },
            "safety_incidents": {
              "type": "string"
            },
            "quality_issues": {
              "type": "string"
            },
            "created_by": {
              "type": "string"
            },
            "created_at": {
              "type": "string",
              "format": "date-time"
            },
            "photos": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PhotoSummary"
              }
            }
          }
        }
      }
    },
    "DailyReportCreate": {
      "type": "object",
      "required": [
        "project_id"
      ],
      "properties": {
        "project_id": {
          "type": "integer",
          "description": "ID of the project"
        },
        "weather_conditions": {
          "type": "string"
        },
        "temperature_high": {
          "type": "number"
        },
        "temperature_low": {
          "type": "number"
        },
        "precipitation": {
          "type": "number"
        },
        "wind_speed": {
          "type": "number"
        },
        "delays": {
          "type": "boolean",
          "default": false
        },
        "delay_description": {
          "type": "string"
        },
        "manpower_count": {
          "type": "integer",
          "default": 0
        },
        "work_performed": {
          "type": "string"
        },
        "materials_received": {
          "type": "string"
        },
        "equipment_used": {
          "type": "string"
        },
        "visitors": {
          "type": "string"
        },
        "safety_incidents": {
          "type": "string"
        },
        "quality_issues": {
          "type": "string"
        }
      }
    },
    "DailyReportResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "report_id": {
          "type": "integer"
        }
      }
    },
    "PhotoSummary": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "file_path": {
          "type": "string"
        },
        "taken_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "SafetyObservationList": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SafetyObservationSummary"
          }
        },
        "count": {
          "type": "integer"
        }
      }
    },
    "SafetyObservationSummary": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "severity": {
          "type": "string"
        },
        "observation_date": {
          "type": "string",
          "format": "date"
        },
        "location": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "observed_by": {
          "type": "string"
        }
      }
    },
    "IncidentList": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/IncidentSummary"
          }
        },
        "count": {
          "type": "integer"
        }
      }
    },
    "IncidentSummary": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "severity": {
          "type": "string"
        },
        "incident_date": {
          "type": "string",
          "format": "date"
        },
        "location": {
          "type": "string"
        },
        "is_osha_recordable": {
          "type": "boolean"
        },
        "is_lost_time": {
          "type": "boolean"
        }
      }
    },
...
  }
}